Split iOSNotificationCenter.cs into several files like iOSNotification.cs, iOSNotificationSettings.cs etc.
