+++ com.unity.mobile.notifications/Runtime/iOS/AuthorizationRequest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b313ac86fae05a14cb6f9c95468075e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
