+++ com.unity.mobile.notifications/Runtime/iOS/iOSNotification.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7ab0873fd8f23f34681825b7642810ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
