+++ com.unity.mobile.notifications/Runtime/iOS/iOSNotificationTriggers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56be22c2d38b781439f9da4ef07e2638
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
